refactor: CharacterAvatar를 Shared/Models로 분리
하면서 BackgroundContainerView도 Shared/Components로 이동
